Adding drawCallbackPixels function
The drawCallbackPixels function is a higher-speed interface to uploading
bulk pixel data, without requiring a single continuous 12kb buffer of
pixel data to be allocated. The callback semantics allow for various
compression algorithms for directly decompressing stored graphics images
into the RGB buffer and other techniques, taking advantage of the delays
the SPI writes cause in the first place.